bugfix: fix capkk playbooks

bugfix: fix capkk playbooks
